+++ CHANGELOG.md
@@ -15,3 +15,12 @@
 
 1. [Fixes](#fixes)
     * Fixed error if the page's content is empty.
+
+# v1.1.0
+## 07/06/2019
+
+1. [New](#new)
+    * Links are now processed correctly according to [Grav rules](https://learn.getgrav.org/16/content/linking).
+
+1. [Fixes](#fixes)
+    * Fixed encoding errors in images ALTs due to Grav's `Excerpts::getExcerptFromHtml` not handling encoding very well.